% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_gbif.R
\name{clean_gbif}
\alias{clean_gbif}
\title{Clean up raw GBIF data}
\usage{
clean_gbif(dat)
}
\arguments{
\item{dat}{a data frame returned by \code{get_gbif()}}
}
\value{
A cleaned data frame containing a single observation for each species. The data frame includes the following columns:
\itemize{
\item{"species"}{The scientific species name}
\item{"lat"}{Degrees latitude of the observation}
\item{"long"}{Degrees longitude of the observation}
\item{"evidence"}{The website for the evidence link}
\item{"date"}{The date that the species was observed (event)}
}
}
\description{
Clean up raw GBIF data
}
\examples{
\dontrun{
clean_gbif(dat)
}
}
\author{
McCrea Cobb \email{mccrea_cobb@fws.gov}
}
