% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{clip_occ}
\alias{clip_occ}
\title{Clip out GBIF observations outside of true refuge boundary}
\usage{
clip_occ(dat, prop, ...)
}
\arguments{
\item{dat}{A data frame of raw species occurrence data for contained within a simplified convex hull of a refuge boundary}

\item{prop}{An \code{sf} multipolygon object of a refuge boundary, returned by \code{get_refuge}.}
}
\value{

}
\description{
Clip out GBIF observations outside of true refuge boundary
}
