% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species.R
\name{get_species}
\alias{get_species}
\title{Get species occurrence data for a refuge}
\usage{
get_species(orgname, ...)
}
\arguments{
\item{orgname}{The organizational name (ORGNAME) for a refuge}

\item{...}{}
}
\value{
A data frame of species occurrence data in a refuge and a map of their locations
}
\description{
Get species occurrence data for a refuge
}
\examples{
\dontrun{
get_species("ARCTIC NATIONAL WILDLIFE REFUGE")
}
}
