% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_wkt}
\alias{get_wkt}
\title{Create a simplified convex hull of the refuge boundary in WKT format}
\usage{
get_wkt(prop, ...)
}
\arguments{
\item{prop}{An \code{sf} multipolygon object of a refuge boundary, returned by \code{get_refuge}.}
}
\value{
A WKT formatted refuge boundary
}
\description{
Create a simplified convex hull of the refuge boundary in WKT format
}
