% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gbif.R
\name{plot_gbif}
\alias{plot_gbif}
\title{Plot species occurrence data on refuges}
\usage{
plot_gbif(prop, dat)
}
\arguments{
\item{prop}{A simple feature (sf) multipolygon object returned \code{get_gbif}}

\item{dat}{A data frame of cleaned species occurrences on refuges returned by \code{clean_gbif}}
}
\value{
A ggplot object
}
\description{
Plot species occurrence data on refuges
}
\examples{
\dontrun{
plot_gbif(dat)
}
}
\author{
McCrea Cobb \email{mccrea@fws.gov}
}
