% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gbif.R
\name{get_gbif}
\alias{get_gbif}
\title{Download species occurrence data from GBIF}
\usage{
get_gbif(orgname = "ARCTIC NATIONAL WILDLIFE REFUGE", pause = 45)
}
\arguments{
\item{pause}{The number of seconds to pause to wait for the data request from GBIF to be ready. Default is 30 secs.}

\item{orgcode}{The org code for desired refuge}
}
\value{
A data frame of species occurrence data from GBIF within the boundary of the desired refuge
}
\description{
Download species occurrence data from GBIF
}
\note{
First time user - Add your GBIF credentials (username, password, email) to an .renviron file using \code{usethis::edit_r_environ(scope = "project")}
\itemize{
\item GBIF_USER = "your GBIF user name"
\item GBIF_PWD = "your GBIF password"
\item GBIF_EMAIL = "your email address"
}
}
\examples{
\dontrun{
get_gbif(orgcode = 75600)
}
}
\author{
McCrea Cobb \email{mccrea@fws.gov}
}
